A bunch of system and quality of life improvements. Added a quadtree system for faster(?) collision detection, added a basic particle system.
This is just a (not so) quick and dirty commit to mark a good milestone with most of the base game done in terms of logic and graphics.
I probably have tons of memory leaks and I've just started working on patching them up as I go along.

Extra notes:
Split a bunch of classes into .h and .cpp files (DAMN YOU, CYCLIC DEPENDENCY!).
Made some core lists static for easier access. I know this isn't really the best idea in terms of cleanliness or multi-threading, but this is just to get things up. I guess the next best thing is to have a singleton master class with most of the essential data in it. Feel free to point me towards better alternatives.
Added a quadtree system to optimize collision detection. This isn't really optimized well right now, with a bunch of if conditions that shouldn't be there and such, it's just an experiment which will hopefully payoff later.
Added some template based functions for easy access to an entity's components. Do you sense some Unity Engine DNA in there? Because I certainly do!
added a particle system. Quite rudimentary, just chuck a bunch of particles in there with a lifetime and it will render them until they die or out of the screen.
Changed the makefile to the capscollective's cpp-starter project makefile. the original raylib one was giving me a bunch of issues regarding not compiling .cpp files, and other weird problems. I fully recommend it if anyone's having trouble with that.