Improved the quadtree system. Fixed a couple of memory leaks
Redesigned the quadtree system to use the standard (Wikipedia) implementation.
Each collider now stores a list of all quadtree nodes it's in, since we're dealing with shapes and not points.

Memory leaks were related to the goal particle effects not destroying their owner along with the quad tree.

Moved "position" to be inside each entity, it now should be destroyed automatically with the entity. If a component requires its own position, it will handle the responsibility of creating and destroying it