A couple of new "systems" : A scene(?) system, and delayed function system, components refactoring, Entity class updates, and more.
Added a scene(? Not really sure) system to be able to have multiple game states (Well, I guess it's a state machine after all...). Looks fancier than a switch statement with a function for each state. I do think this is a lot more expandable, with each state/scene being able to have its own sub-state.

Added an event(?) queue for delayed function calls. You can see a simple use of this inside Paddle.cpp where I set a boolean that's responsible for rendering a segment to true, then I reset it after a delay.
I tried making this system to be able to pass any void function with a variable number of arguments to provide flexibility, not really sure how I can implement functions with different return types, but that's not really an issue right now.

Added the ability to enable or disable specific components or all components on an entity. Might be useful to someone (or me in the future)

Added OnCollisionStay(), this event will fire for every frame 2 colliders are colliding.

Moved to using CapCollective's raylib-cpp-starter Makefile, had to do some restructuring for that.

Refactored some classes (Ball, Particle, Paddle).