Fix bug with `MultiIterator` where it would iterate past the end.
This was due to `_begin` and `_end` having different end iterators. The former
had the end iterator of the component vector (whether it was valid or
not), while the latter had an iterator to the last valid component. This
would cause issues when comparing the current iterator (copied from
`_begin`) to `_end`