Fix `getComponentIterAtOffset` returning an iterator to a temporary component vector
Turns out that `auto` and `auto&` bind to copies and references
respectively. This caused us to copy the component vector onto the
stack and return an iterator pointing to the copy on the stack, not the
one held in the ECS memory.

Later on when we would try to access this iterator, it would point to
garbage memory, causing a SIGSEGV.